+++ app/Http/Controllers/LineWebhookController.php
@@ -1,34 +1,57 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Http\Controllers;
 
-use App\Contracts\LineMessagingServiceInterface;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
-use Exception;
-use Illuminate\Http\JsonResponse;
+use App\Services\Messaging\DTOs\LineWebhookData;
+use Illuminate\Support\Facades\Log;
 
-final class LineWebhookController extends Controller
+class LineWebhookController extends Controller
 {
-    public function __construct(
-        private readonly LineMessagingServiceInterface $lineMessagingService
-    ) {}
-
-    /**
-     * LINEからのWebhookリクエストを処理する
-     *
-     * @param Request $request
-     * @return Response|JsonResponse
-     */
-    public function handle(Request $request): Response|JsonResponse
+    public function handle(Request $request)
     {
-        try {
-            $this->lineMessagingService->handleWebhook($request);
-            return response()->noContent();
-        } catch (Exception $e) {
-            return response()->json(['error' => 'Invalid signature'], 401);
+        // LineWebhookDataオブジェクトを作成
+        $webhookData = LineWebhookData::fromRequest($request);
+
+        // 署名の検証（必要に応じて実装）
+        if (!$this->isValidSignature($webhookData->signature, $webhookData->content)) {
+            return response()->json(['message' => 'Invalid signature'], 400);
         }
+
+        // イベントごとに処理を実行
+        foreach ($webhookData->events as $event) {
+            $this->handleEvent($event);
+        }
+
+        return response()->json(['message' => 'OK'], 200);
+    }
+
+    private function isValidSignature($signature, $content)
+    {
+        // 署名の検証ロジックを実装
+        // 例: hash_hmac('sha256', $content, $channelSecret) === $signature
+        return true;
+    }
+
+    private function handleEvent(array $event)
+    {
+        $eventType = LineWebhookData::getEventType($event);
+        $userId = LineWebhookData::getUserId($event);
+
+        switch ($eventType) {
+            case 'message':
+                $this->handleMessageEvent($event, $userId);
+                break;
+            // 他のイベントタイプに対する処理を追加
+            default:
+                Log::info('Unhandled event type: ' . $eventType);
+                break;
+        }
+    }
+
+    private function handleMessageEvent(array $event, ?string $userId)
+    {
+        // メッセージイベントの処理ロジックを実装
+        Log::info('Received message event from user: ' . $userId);
     }
 }
